# swagger.yaml
openapi: 3.0.3
info:
  title: Blood Management System
  version: 1.0.0
  description:
    "The Blood Donation Management System backend is designed to provide
    REST APIs for blood donation activities including donor information management,
    blood request processing, and role-based access control. The system will support
    multiple microservices with appropriate authentication and authorization mechanisms.
    "
  contact:
    email: mr.monirul.dev@gmail.com
    name: Monirul Islam
    url: http://github.com/codemonkmi
servers:
  - url: http://localhost:4000/api/v1
    description: dev server
...
paths:
...
  /requests:
    get:
      security:
        - BearerAuth: []
      tags:
        - Donation
      summary: Get All Requests
      parameters:
        - $ref: "#/components/parameters/pageParam"
        - $ref: "#/components/parameters/limitParam"
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: array
                    items:
                      allOf:
                        - $ref: "#/components/schemas/DonationRequested"
                        - $ref: "#/components/schemas/Timestamp"
                        - type: object
                          properties:
                            link:
                              type: string
                              example: /request/2
                  pagination:
                    $ref: "#/components/schemas/PaginationMetadata"
                  link:
                    type: object
                    properties:
                      self:
                        type: string
                        example: /request
        400:
          $ref: "#/components/responses/400"
        401:
          $ref: "#/components/responses/401"
        403:
          $ref: "#/components/responses/403"
        500:
          $ref: "#/components/responses/500"
      description: Performs GET operation on /requests

    post:
      tags:
        - Donation
      summary: Submit Blood Request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BloodRequestDTO"
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    allOf:
                      - type: object
                        properties:
                          id:
                            type: string
                            example: req_001
                          status:
                            $ref: "#/components/schemas/DonationStatus"

                      - type: object
                        properties:
                          id:
                            type: string
                            example: 2
                          name:
                            type: string
                            example: Monirul Islam
                          phone:
                            type: string
                            example: "01234567890"
                          address:
                            type: string
                            example: Dhaka, Bangladesh
                          date:
                            type: string
                            format: date-time
                          blood:
                            $ref: "#/components/schemas/BloodType"
                          description:
                            type: string
                            example: "Blood tech is a general term that can refer to several different, but related, professions in the healthcare field"
                          status:
                            $ref: "#/components/schemas/DonationStatus"
                      - $ref: "#/components/schemas/Timestamp"
                  link:
                    type: object
                    properties:
                      self:
                        type: string
                        example: /request
        400:
          $ref: "#/components/responses/400"
        500:
          $ref: "#/components/responses/500"
      description: Performs POST operation on /requests
  /requests/{id}:
    get:
      security:
        - BearerAuth: []
      tags:
        - Donation
      summary: Get Request
      parameters:
        - $ref: "#/components/parameters/idParam"
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                    example: Request Successful
                  data:
                    allOf:
                      - $ref: "#/components/schemas/DonationRequested"
                      - $ref: "#/components/schemas/Timestamp"
                  link:
                    type: object
                    properties:
                      self:
                        type: string
                        example: /request/2
        400:
          $ref: "#/components/responses/400"
        401:
          $ref: "#/components/responses/401"
        403:
          $ref: "#/components/responses/403"
        404:
          $ref: "#/components/responses/404"
        500:
          $ref: "#/components/responses/500"
      description: Performs GET operation on /requests/{id}

    patch:
      security:
        - BearerAuth: []
      tags:
        - Donation
      summary: Update Request
      parameters:
        - $ref: "#/components/parameters/idParam"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BloodRequestUpdateDTO"
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                    example: Request updated Successful
                  data:
                    allOf:
                      - $ref: "#/components/schemas/DonationRequested"
                      - $ref: "#/components/schemas/Timestamp"
                  link:
                    type: object
                    properties:
                      self:
                        type: string
                        example: /request/2
        400:
          $ref: "#/components/responses/400"
        401:
          $ref: "#/components/responses/401"
        403:
          $ref: "#/components/responses/403"
        404:
          $ref: "#/components/responses/404"
        500:
          $ref: "#/components/responses/500"
      description: Performs PATCH operation on /requests/{id}
...
  /history:
    get:
      security:
        - BearerAuth: []
      tags:
        - Donation
      summary: Get All History
      parameters:
        - $ref: "#/components/parameters/pageParam"
        - $ref: "#/components/parameters/limitParam"
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: array
                    items:
                      allOf:
                        - $ref: "#/components/schemas/DonationHistory"
                        - $ref: "#/components/schemas/Timestamp"
                        - type: object
                          properties:
                            link:
                              type: string
                              example: /history/2
                  pagination:
                    $ref: "#/components/schemas/PaginationMetadata"
                  link:
                    type: object
                    properties:
                      self:
                        type: string
                        example: /history
                      prev:
                        type: string
                        example: /history?page=1
                      next:
                        type: string
                        example: /history?page=3
        400:
          $ref: "#/components/responses/400"
        401:
          $ref: "#/components/responses/401"
        403:
          $ref: "#/components/responses/403"
        404:
          $ref: "#/components/responses/404"
        500:
          $ref: "#/components/responses/500"
      description: Performs GET operation on /history

    post:
      security:
        - BearerAuth: []
      tags:
        - Donation
      summary: Create History
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HistoryCreateDTO"
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    allOf:
                      - $ref: "#/components/schemas/DonationHistory"
                      - $ref: "#/components/schemas/Timestamp"
                  links:
                    type: object
                    properties:
                      self:
                        type: string
                        format: url
                        example: /history/2
        400:
          $ref: "#/components/responses/400"
        401:
          $ref: "#/components/responses/401"
        403:
          $ref: "#/components/responses/403"
        404:
          $ref: "#/components/responses/404"
        500:
          $ref: "#/components/responses/500"
      description: Performs POST operation on /history
  /history/{id}:
    get:
      security:
        - BearerAuth: []
      tags:
        - Donation
      summary: Get All History
      parameters:
        - $ref: "#/components/parameters/idParam"
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    allOf:
                      - $ref: "#/components/schemas/DonationHistory"
                      - $ref: "#/components/schemas/Timestamp"
                  link:
                    type: object
                    properties:
                      self:
                        type: string
                        example: /history/2
        400:
          $ref: "#/components/responses/400"
        401:
          $ref: "#/components/responses/401"
        403:
          $ref: "#/components/responses/403"
        404:
          $ref: "#/components/responses/404"
        500:
          $ref: "#/components/responses/500"
      description: Performs GET operation on /history/{id}
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    BloodType:
      type: string
      enum:
        - A_POSITIVE
        - A_NEGATIVE
        - B_POSITIVE
        - B_NEGATIVE
        - AB_POSITIVE
        - AB_NEGATIVE
        - O_POSITIVE
        - O_NEGATIVE
    DonationHistory:
      type: object
      properties:
        id:
          type: string
          example: 2
        donor:
          type: string
          description: User ID of the donor
          example: 1
        request:
          type: string
          description: User ID of the requester
          example: 2
        status:
          $ref: "#/components/schemas/DonationStatus"
        message:
          type: string
          example: Status updated to progress
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - donor
        - request
        - status
        - message
    DonationRequested:
      type: object
      properties:
        id:
          type: string
          example: 2
        name:
          type: string
          example: Monirul Islam
        phone:
          type: string
          example: "01234567890"
        address:
          type: string
          example: Dhaka, Bangladesh
        date:
          type: string
          format: date-time
        blood:
          $ref: "#/components/schemas/BloodType"
        description:
          type: string
          example: "Blood tech is a general term that can refer to several different, but related, professions in the healthcare field"
        status:
          $ref: "#/components/schemas/DonationStatus"
        metadata:
          type: object
        donor:
          type: string
          example: 1
        details:
          type: string
          example: "Blood tech is a general term that can refer to several different, but related, professions in the healthcare field."

      required:
        - id
        - name
        - phone
        - address
        - date
        - blood
        - status
    DonationStatus:
      type: string
      enum:
        - pending
        - verified
        - progress
        - ready
        - hold
        - completed
...
    PaginationMetadata:
      type: object
      properties:
        page:
          type: integer
          example: 2
        limit:
          type: integer
          example: 10
        totalItems:
          type: integer
          example: 100
        totalPages:
          type: integer
          example: 10
        next:
          type: integer
          nullable: true
          example: 3
        prev:
          type: integer
          nullable: true
          example: 1

      required:
        - page
        - limit
        - totalItems
        - totalPages
...
    Timestamp:
      type: object
      properties:
        created_at:
          type: string
          format: date-time
          example: "2024-06-20T12:34:56Z"
        updated_at:
          type: string
          format: date-time
          example: "2024-06-21T08:45:00Z"
      required:
        - created_at
        - updated_at
...
    HistoryCreateDTO:
      type: object
      required:
        - donor
        - request
        - status
        - message
      properties:
        donor:
          type: string
          example: 1
        request:
          type: string
          example: 1
        status:
          $ref: "#/components/schemas/DonationStatus"
        message:
          $ref: "#/components/schemas/BloodType"
    BloodRequestDTO:
      type: object
      required:
        - name
        - phone
        - address
        - date
        - blood
        - description
      properties:
        name:
          type: string
          example: John Doe
        phone:
          type: string
          example: 01234567890
        address:
          type: string
          example: Dhaka Bangladesh
        date:
          type: string
          format: date-time
        blood:
          $ref: "#/components/schemas/BloodType"
        description:
          type: string
          example: Need for pregnancy operations
    BloodRequestUpdateDTO:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        address:
          type: string
        blood:
          $ref: "#/components/schemas/BloodType"
        date:
          type: string
          format: date-time
        description:
          type: string
        status:
          $ref: "#/components/schemas/DonationStatus"

      required:
        - name
        - phone
        - address
        - blood
        - date
        - description
        - status
...
  parameters:
    pageParam:
      name: page
      in: query
      description: The page number to retrieve
      required: false
      schema:
        type: integer
        default: 2
        minimum: 1
    limitParam:
      in: query
      name: limit
      description: Number of items per page
      required: false
      schema:
        type: integer
        default: 10
        minimum: 1
        maximum: 100
...
    idParam:
      in: path
      name: id
      description: Resource Id
      schema:
        type: string
        example: 2
      required: true
  responses:
    400:
      description: Bad request
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                format: boolean
                example: false
              message:
                type: string
                format: string
                example: Bad request
              data:
                type: array
                items:
                  type: object
                  properties:
                    code:
                      type: string
                      format: string
                      example: too_small
                    pathname:
                      type: array
                      items:
                        type: string
                        format: string
                        example: id
                    message:
                      type: string
                      format: string
                      example: Invalid id
    401:
      description: Unauthorized
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                format: boolean
                example: false
              message:
                type: string
                format: string
                example: Unauthorized
    403:
      description: Unauthorized
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                format: boolean
                example: false
              message:
                type: string
                format: string
                example: You are not authorized to perform this actions
    404:
      description: Not Found
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                format: boolean
                example: false
              message:
                type: string
                format: string
                example: Your requested resource not found!
    500:
      description: Internal Server Error
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                format: boolean
                example: false
              message:
                type: string
                format: string
                example: We sorry for the inconvenience. We will fix this asap
...